fix: move OperatorType type on Operators.model.ts
