+++ src/components/Operation/Operation.tsx
@@ -1,13 +1,14 @@
 import React from "react";
+import {OperatorionsType} from "../../models/Operations";
 
 interface OperationProps {
-  operation: string;
-  clickOperation: () => void;
+  operation: OperatorionsType;
+  clickOperation: (operations: OperatorionsType) => void;
 }
 
 const Operation = ({operation, clickOperation}: OperationProps) => {
   return (
-    <button className="operation" onClick={clickOperation}>
+    <button className="operation" onClick={() => clickOperation(operation)}>
       {operation}
     </button>
   );
