+++ src/constants/Calcurator.ts
@@ -5,3 +5,5 @@ export const OPERATIONS = {
   plus: "+",
   equalSign: "=",
 } as const;
+
+export const NUMERICS = [9, 8, 7, 6, 5, 4, 3, 2, 1, 0] as const;
