+++ src/models/Operations.ts
@@ -1,0 +1,3 @@
+import {OPERATIONS} from "../constants/Calcurator";
+
+export type OperatorionsType = typeof OPERATIONS[keyof typeof OPERATIONS];
